refactor: extract CarsEditorPanel for profile and trip flows.
Share cars CRUD between ProfileCars settings page and TripCarsModal to remove duplicated editor logic.
