+++ src/components/sections/ProfileCars.view.test.js
@@ -3,11 +3,13 @@ import path from 'node:path';
 import { describe, expect, it } from 'vitest';
 
 const viewPath = path.join(__dirname, 'ProfileCars.vue');
+const panelPath = path.join(__dirname, '../elements/CarsEditorPanel.vue');
 const formPath = path.join(__dirname, '../elements/CarForm.vue');
 
 describe('ProfileCars CRUD section', () => {
     it('defines profile cars view and car form components', () => {
         expect(fs.existsSync(viewPath)).toBe(true);
+        expect(fs.existsSync(panelPath)).toBe(true);
         expect(fs.existsSync(formPath)).toBe(true);
     });
 
@@ -16,37 +18,13 @@ describe('ProfileCars CRUD section', () => {
 
         expect(viewSource).toContain('CarsEditorPanel');
         expect(viewSource).not.toContain('CarForm');
+        expect(viewSource).toContain(':prompt-incomplete-on-load="true"');
     });
 
-    it('lists cars with add, edit and delete actions', () => {
+    it('shows autos heading on profile settings page', () => {
         const viewSource = fs.readFileSync(viewPath, 'utf8');
 
         expect(viewSource).toContain("$t('autos')");
-        expect(viewSource).toContain("$t('agregarAuto')");
-        expect(viewSource).toContain("$t('editarAuto')");
-        expect(viewSource).toContain('carDisplayLabel');
-        expect(viewSource).toContain('openEditCar');
-        expect(viewSource).toContain('deleteCar');
-        expect(viewSource).toContain('openAddCar');
-        expect(viewSource).toContain('btn-danger');
-        expect(viewSource).toContain('fa-pencil');
-        expect(viewSource).toContain('fa-trash-o');
-        expect(viewSource).toContain('fa-plus');
-    });
-
-    it('requires complete marca, modelo, año and color before saving', () => {
-        const viewSource = fs.readFileSync(viewPath, 'utf8');
-
-        expect(viewSource).toContain('isCarFormComplete');
-        expect(viewSource).toContain('carFormMissingFieldKeys');
-        expect(viewSource).toContain('autosDatosIncompletos');
-    });
-
-    it('prompts user to complete each incomplete saved car', () => {
-        const viewSource = fs.readFileSync(viewPath, 'utf8');
-
-        expect(viewSource).toContain('carsNeedingCompletion');
-        expect(viewSource).toContain('autosCompletarPendientes');
     });
 
     it('uses readable text colors and mobile-only page heading', () => {
@@ -55,7 +33,6 @@ describe('ProfileCars CRUD section', () => {
         expect(viewSource).toContain('visible-xs-block');
         expect(viewSource).toContain('color: #036686');
         expect(viewSource).toContain('var(--main-font-color');
-        expect(viewSource).toContain('profile-cars__label');
     });
 });
